Extract aidev_lib/budget.py (#17 Stage 2)
Move budget, cost estimation, token accounting, and ledger helpers
out of aidev.py into a focused module. aidev.py shrinks 1328 → 1130
(-198 lines); aidev_lib/budget.py is 247 lines.

Moved:
- DEFAULT_BUDGET (loaded from shared/budget-defaults.json)
- estimate_cost, model_token_price, estimate_token_cost
- estimated_output_tokens_for_reasoning, estimate_tokens
- model_context_limit, extract_cli_token_usage, build_usage_report
- session_estimated_spend, daily_usage, budget_check, append_ledger

The four functions that touch the on-disk ledger now take a budget_dir
parameter. aidev.py keeps thin re-export shims that pass BUDGET_DIR so
existing callers and pytest cases continue to import these names from
aidev directly.

Verified:
- python -m pytest tests/ → 71 passed
- cd desktop && npm run smoke → all 71 unit tests pass

Refs #17.
